(trivial-ldap:search): Add around method to check arguments for sensible defaults.
What happens now is when a search is done, we look up the LDAP object to
use as well as look up the search filter to use by calling COMPUTE-LDAP
and COMPUTE-FILTER respectively.

The key part here is COMPUTE-FILTER and COMPUTE-LDAP are both generic
functions. This means a custom filter and or a custom ldap object can be
added with no issues. This is up to and including a custom object
representing an LDAP query, and or a custom object representing an LDAP
connection.